refactor: move non-worker spec/mtp files from runtime to framework/speculative.
