+++ tests/core/framework/CMakeLists.txt
@@ -12,6 +12,7 @@ add_subdirectory(parallel_state)
 add_subdirectory(prefix_cache)
 add_subdirectory(request)
 add_subdirectory(sampling)
+add_subdirectory(speculative)
 add_subdirectory(state_dict)
 add_subdirectory(tokenizer)
