+++ tests/core/runtime/CMakeLists.txt
@@ -3,28 +3,6 @@ include(cc_test)
 if(USE_NPU)
   cc_test(
     NAME
-      mtp_async_input_builder_test
-    SRCS
-      mtp_async_input_builder_test.cpp
-    DEPS
-      :runtime
-      :models
-      GTest::gtest_main
-      torch
-      torch_npu
-      pybind11::embed
-  )
-  target_link_libraries(mtp_async_input_builder_test
-                        PRIVATE
-                        torch_python
-                        Python::Python
-                        ascendcl
-                        hccl
-                        c_sec
-                        nnopbase)
-
-  cc_test(
-    NAME
       dspark_confidence_head_test
     SRCS
       dspark_confidence_head_test.cpp
@@ -111,62 +89,6 @@ if(USE_NPU)
     "${CMAKE_BINARY_DIR}/third_party/spdlog/libspdlog.a"
     "-Wl,--no-whole-archive")
 endif()
-
-cc_test(
-  NAME
-    spec_input_builder_test
-  SRCS
-    spec_input_builder_test.cpp
-    "${PROJECT_SOURCE_DIR}/xllm/core/runtime/spec_input_builder.cpp"
-  DEPS
-    :multimodal
-    :parallel_state
-    torch
-    glog::glog
-    proto::xllm_proto
-    GTest::gtest_main
-)
-
-cc_test(
-  NAME
-    mtp_async_state_test
-  SRCS
-    mtp_async_state_test.cpp
-    "${PROJECT_SOURCE_DIR}/xllm/core/runtime/mtp_async_state.cpp"
-  DEPS
-    torch
-    glog::glog
-    GTest::gtest_main
-)
-if(EXISTS "$ENV{PYTORCH_INSTALL_PATH}/../torch.libs")
-  target_link_options(mtp_async_state_test PRIVATE
-    "-Wl,-rpath-link,$ENV{PYTORCH_INSTALL_PATH}/../torch.libs")
-endif()
-
-cc_test(
-  NAME
-    adaptive_speculative_controller_test
-  SRCS
-    adaptive_speculative_controller_test.cpp
-    "${PROJECT_SOURCE_DIR}/xllm/core/framework/speculative/adaptive_speculative_controller.cpp"
-    "${PROJECT_SOURCE_DIR}/xllm/core/framework/speculative/speculative_profile_registry.cpp"
-  DEPS
-    torch
-    glog::glog
-    proto::xllm_proto
-    GTest::gtest_main
-)
-
-cc_test(
-  NAME
-    speculative_profile_registry_test
-  SRCS
-    speculative_profile_registry_test.cpp
-    "${PROJECT_SOURCE_DIR}/xllm/core/framework/speculative/speculative_profile_registry.cpp"
-  DEPS
-    glog::glog
-    GTest::gtest_main
-)
 
 if(USE_MLU)
   cc_test(
