+++ xllm/core/framework/kv_cache/CMakeLists.txt
@@ -32,7 +32,6 @@ cc_library(
     kv_cache
   HDRS
     deepseek_v4_kv_cache_impl.h
-    embedding_cache.h
     indexed_kv_cache_impl.h
     kv_cache.h
     kv_cache_impl.h
@@ -45,7 +44,6 @@ cc_library(
     quantized_kv_cache_impl.h
   SRCS
     deepseek_v4_kv_cache_impl.cpp
-    embedding_cache.cpp
     indexed_kv_cache_impl.cpp
     kv_cache.cpp
     kv_cache_impl.cpp
