+++ xllm/core/framework/speculative/embedding_cache.h
