+++ xllm/core/framework/speculative/spec_input_builder.h
