+++ xllm/core/runtime/mtp_worker_impl.cpp
@@ -46,11 +46,11 @@ limitations under the License.
 #include "core/layers/common/expanded_decode_metadata_builder.h"
 #endif
 #include "core/framework/speculative/adaptive_pruning_helpers.h"
+#include "core/framework/speculative/mtp_async_input_builder.h"
+#include "core/framework/speculative/mtp_async_state.h"
+#include "core/framework/speculative/spec_input_builder.h"
 #include "core/framework/speculative/speculative_profile_registry.h"
 #include "core/layers/common/dsa_topk_share_plan.h"
-#include "core/runtime/mtp_async_input_builder.h"
-#include "core/runtime/mtp_async_state.h"
-#include "spec_input_builder.h"
 #include "util/pretty_print.h"
 #include "util/slice.h"
 #include "util/timer.h"
